% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{all_combos}
\alias{all_combos}
\title{All combinations}
\usage{
all_combos(x, min = 1, max = length(x))
}
\arguments{
\item{x}{Vector to get combinations of}

\item{min}{Minimum number of items in a combination}

\item{max}{Maximum number of items in a combination}
}
\value{
Nested named list of combinations. The first level of the list is the
combination lengths and the second level is the combinations themselves
}
\description{
Compute all possible combinations of items in a vector
}
\examples{
all_combos(1:3)

# Single level list of combinations
unlist(all_combos(1:3), recursive = FALSE)
}
