% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cats.R
\name{cat_function}
\alias{cat_function}
\title{A Cat Function}
\usage{
cat_function(love = TRUE)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats. It comes from
Hilary Parker's \href{https://hilaryparker.com/2014/04/29/writing-an-r-package-from-scratch/}{"Writing an R package from scratch"}.
}
\examples{
cat_function()
}
