% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rounding.R
\name{chrRound}
\alias{chrRound}
\title{Round a number to a character, preserving extra 0's}
\usage{
chrRound(x, digits = 1)
}
\arguments{
\item{x}{Number to round.}

\item{digits}{Number of digits past the decimal point to keep.}
}
\value{
A vector of character strings.
}
\description{
Round a number to a character, preserving extra 0's
}
\details{
Uses \code{\link[base]{sprintf}} to round a number, keeping extra 0's. Useful
for displaying values. Based on \code{\link[broman]{myround}}.
}
\examples{
chrRound(51.01, 3)
chrRound(0.199, 2)
}
\seealso{
\code{\link[base]{round}}, \code{\link[base]{sprintf}}
}
\keyword{utilities}

