% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{combine_sets}
\alias{combine_sets}
\title{Combine sets}
\usage{
combine_sets(sets, method = c("union", "intersect"), sort = TRUE)
}
\arguments{
\item{sets}{List of vectors}

\item{method}{Method for combining sets}

\item{sort}{Whether to sort the results}
}
\value{
Named list of combined vectors
}
\description{
Compute all combinations of a list of vectors using a selected set method
}
\examples{
x <- list(A = c(1, 2, 3), B = c(2, 3, 4), C = c(3, 4, 5))
combine_sets(x, method = "union")
combine_sets(x, method = "intersect")
}
