% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sets.R
\name{combineSets}
\alias{combineSets}
\title{Combine Sets}
\usage{
combineSets(set.list, method = c("union", "intersect"), sort = TRUE)
}
\arguments{
\item{set.list}{List of vectors to combine}

\item{method}{Method to combine sets, can be "union" or "intersect"
abbreviations accepted)}

\item{sort}{Whether to sort the final sets, default is TRUE}
}
\value{
List containing combined sets
}
\description{
Compute combination of a list of vectors using a selected set method
}

