% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{defaultChunkOpts}
\alias{defaultChunkOpts}
\title{Default Chunk Options}
\usage{
defaultChunkOpts()
}
\description{
Set default chunk options for knitr
}
\details{
Sets my preferred options for how knitr displays code and figures
}
\examples{
defaultChunkOpts()

}
\seealso{
\code{\link[knitr]{opts_chunk}}
}
\keyword{rmarkdown}

