% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stats.R
\name{finiteMean}
\alias{finiteMean}
\title{Finite Mean}
\usage{
finiteMean(x)
}
\arguments{
\item{x}{Numeric vector to calculate mean}
}
\value{
Mean value
}
\description{
Calculate mean of a vector excluding infinite and NA values
}
\examples{
mean(c(1, 2, 4, 2, Inf))
finiteMean(c(1, 2, 4, 2, Inf))
mean(c(1, 2, 4, 2, Inf, NA))
finiteMean(c(1, 2, 4, 2, Inf, NA))
}

