% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{intersect_list}
\alias{intersect_list}
\title{Intersect list}
\usage{
intersect_list(x, sort = TRUE)
}
\arguments{
\item{x}{List of vectors to intersect}

\item{sort}{Whether to sort the intersection}
}
\value{
Vector containing the intersection of all vectors in \code{x}
}
\description{
Find the intersection of a list of vectors
}
\examples{
x <- list(A = c(1, 2, 3), B = c(2, 3, 4), C = c(3, 4, 5))
intersect_list(x)
}
