% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sets.R
\name{listIntersect}
\alias{listIntersect}
\title{List Intersect}
\usage{
listIntersect(set.list, sort = TRUE)
}
\arguments{
\item{set.list}{List of vectors to intersect}

\item{sort}{Whether to sort the final set, default is TRUE}
}
\value{
Vector containing the intersection
}
\description{
Find the intersect of a list of vectors
}
\examples{
l <- list(Set1 = sample(20, 10), Set2 = sample(20, 10),
          Set3 = sample(20, 10), Set4 = sample(20, 10))
listIntersect(l)
}

