% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sets.R
\name{listSetdiff}
\alias{listSetdiff}
\title{List Setdiff}
\usage{
listSetdiff(set.list1, set.list2, sort = TRUE)
}
\arguments{
\item{set.list1}{The first list of vectors to compare}

\item{set.list2}{The second list of vectors to compare}

\item{sort}{Whether to sort the final set, default is TRUE}
}
\value{
List containing the items in set.list1 that are not in set.list2
}
\description{
Find the items that are in one list of vector but not another
}

