% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{name_seq}
\alias{name_seq}
\title{Name sequentially}
\usage{
name_seq(x, prefix = "")
}
\arguments{
\item{x}{Object to name}

\item{prefix}{Prefix string added to each new name}
}
\value{
Object with names
}
\description{
Sequentially name the items in an object. If names already exist they will be
kept but empty or \code{NA} names will be replaced.
}
\examples{
name_seq(1:3)
name_seq(c(A = 1, 2, 3), prefix = "Item")
}
