% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{overlapSets}
\alias{overlapSets}
\title{Overlap Sets}
\usage{
overlapSets(set.list, sort = TRUE)
}
\arguments{
\item{set.list}{List of vectors to overlap}

\item{sort}{Whether to sort the final sets, default is TRUE}
}
\value{
List of sets corresponding to Venn regions
}
\description{
Takes a list of vectors and returns the sets corresponding to the overlapping
regions in a Venn diagram
}
\examples{
l <- list(Set1 = sample(20, 10), Set2 = sample(20, 10),
          Set3 = sample(20, 10), Set4 = sample(20, 10))
overlapSets(l)

}

