% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.R
\name{scale_mean}
\alias{scale_mean}
\title{Scale Mean}
\usage{
scale_mean(x)
}
\arguments{
\item{x}{Numeric vector to scale}
}
\value{
Scaled vector
}
\description{
Mean scale a vector
}
\examples{
scale_mean(1:10)
scale_mean(-5:10)

}
\references{
https://en.wikipedia.org/wiki/Feature_scaling#Mean_normalization
}
