% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.R
\name{scale_minmax}
\alias{scale_minmax}
\title{Scale Min-Max}
\usage{
scale_minmax(x, from = range(x, na.rm = TRUE, finite = TRUE))
}
\arguments{
\item{x}{Continuous vector to scale}

\item{from}{Input range (vector of length two). If not given is calculated
from the range for \code{x}}
}
\value{
Scaled vector
}
\description{
Scale a vector to be between 0 and 1
}
\examples{
x <- 0:10
scale_minmax(x)
scale_minmax(x, from = c(-10, 10))

}
\seealso{
scales::rescale
}
