% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.R
\name{scale_unit}
\alias{scale_unit}
\title{Scale Unit Length}
\usage{
scale_unit(x)
}
\arguments{
\item{x}{Numeric vector to scale}
}
\value{
Scaled vector
}
\description{
Scale a vector to unit length
}
\examples{
scale_unit(1:10)
scale_unit(-5:10)

}
\references{
https://en.wikipedia.org/wiki/Feature_scaling#Scaling_to_unit_length
}
