% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{scaleRange}
\alias{scaleRange}
\title{Scale Range}
\usage{
scaleRange(x, newMin = 0, newMax = 1, ...)
}
\arguments{
\item{x}{Numeric vector to scale}

\item{newMin}{New minimum value}

\item{newMax}{New maximum value}

\item{...}{Extra parameter passed to min and max (eg na.rm = TRUE)}
}
\value{
Scaled vector
}
\description{
Scale a vector to be between a new range
}
\examples{
x <- 0:10
scaleRange(x)
scaleRange(x, -2, 2)
scaleRange(c(x, NA))
scaleRange(c(x, NA), na.rm = TRUE)

}
\references{
https://stackoverflow.com/questions/5468280/scale-a-series-between-two-points
}
\seealso{
scales::rescale
}

