% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{setdiff_lists}
\alias{setdiff_lists}
\title{Setdiff lists}
\usage{
setdiff_lists(x, y, sort = TRUE)
}
\arguments{
\item{x}{List of vectors}

\item{y}{Another list of vectors}

\item{sort}{Whether to sort the difference}
}
\value{
Vector containing difference between two lists of vectors
}
\description{
Find the set of items that are present in the intersection of one list of
vectors but not present in the union of another list of vectors
}
\examples{
x <- list(A = c(1, 2, 3), B = c(2, 3, 4), C = c(3, 4, 5))
y <- list(D = c(1, 2), E = c(4, 5))
setdiff_lists(x, y)
}
