% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_lazappi}
\alias{theme_lazappi}
\title{Theme lazappi}
\usage{
theme_lazappi(base.size = 12, base.family = "")
}
\arguments{
\item{base.size}{Base font size}

\item{base.family}{Base font family}
}
\value{
ggplot2 theme object
}
\description{
My ggplot2 theme
}

