% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{union_list}
\alias{union_list}
\title{Union list}
\usage{
union_list(x, sort = TRUE)
}
\arguments{
\item{x}{List of vectors to union}

\item{sort}{Whether to sort the union}
}
\value{
Vector containing the union of all vectors in \code{x}
}
\description{
Find the union of a list of vectors
}
\examples{
x <- list(A = c(1, 2, 3), B = c(2, 3, 4), C = c(3, 4, 5))
union_list(x)
}
