% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{venn_sets}
\alias{venn_sets}
\title{Venn sets}
\usage{
venn_sets(sets, sort = TRUE)
}
\arguments{
\item{sets}{List of vectors}

\item{sort}{Whether to sort the overlaps}
}
\value{
Named list of overlap vectors
}
\description{
Takes a list of vectors and returns the sets corresponding to the overlapping
regions in a Venn diagram
}
\examples{
x <- list(A = c(1, 2, 3), B = c(2, 3, 4), C = c(3, 4, 5))
venn_sets(x)
}
