% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{combineSets}
\alias{combineSets}
\title{Combine Sets}
\usage{
combineSets(set.list, method = c("union", "intersect"), sort = TRUE)
}
\arguments{
\item{set.list}{List of vectors to combine}

\item{method}{Method to combine sets, can be "union" or "intersect"
abbreviations accepted)}

\item{sort}{Whether to sort the final sets, default is TRUE}
}
\value{
List containing combined sets
}
\description{
Compute combination of a list of vectors using a selected set method
}
\examples{
l <- list(Set1 = sample(20, 10), Set2 = sample(20, 10),
          Set3 = sample(20, 10), Set4 = sample(20, 10))
combineSets(l, "union")
combineSets(l, "intersect")

}

