% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sets.R
\name{listUnion}
\alias{listUnion}
\title{List Union}
\usage{
listUnion(set.list, sort = TRUE)
}
\arguments{
\item{set.list}{List of vectors to union}

\item{sort}{Whether to sort the final set, default is TRUE}
}
\value{
Vector containing the union
}
\description{
Find the union of a list of vectors
}
\examples{
l <- list(Set1 = sample(20, 10), Set2 = sample(20, 10),
          Set3 = sample(20, 10), Set4 = sample(20, 10))
listUnion(l)
}

